requirements.txt
absl-py==0.15.0
astor==0.8.0
astunparse==1.6.3
biwrap==0.1.6
cached-property==1.5.2
cachetools==4.1.0
certifi==2019.6.16
chardet==3.0.4
clang==5.0
cloudpickle==1.4.1
cycler==0.10.0
decorator==4.4.1
dm-tree==0.1.6
et-xmlfile==1.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.1.0
idna==2.8
jdcal==1.4.1
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mock==4.0.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.1.0.25
openpyxl==2.6.2
opt-einsum==3.3.0
panda==0.3.1
pandas==0.25.0
Pillow==6.1.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.7.0
tensorflow-gpu==2.6.0
tensorflow-plot==0.3.2
tensorflow-probability==0.14.0
termcolor==1.1.0
Theano==1.0.4
tqdm==4.32.2
typing-extensions==3.7.4.3
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.12.1
